[Comgr] Port get_data_isa_name_test from CTest to LIT
Ports get_data_isa_name_test to LIT and splits into two separate tests, isa-enumeration and get-data-isa-name.
