+++ amd/comgr/test-lit/CMakeLists.txt
@@ -54,5 +54,7 @@ add_comgr_lit_binary(status-string c)
 add_comgr_lit_binary(data-action c)
 add_comgr_lit_binary(lookup-code-object c)
 add_comgr_lit_binary(hotswap-rewrite c)
+add_comgr_lit_binary(isa-enumeration c)
+add_comgr_lit_binary(get-data-isa-name c)
 
 add_dependencies(check-comgr test-lit)
