+++ amd/comgr/test-lit/isa-enumeration.c
@@ -1,0 +1,59 @@
+// COM: Test Comgr get_isa_count() and amd_comgr_get_isa_name() API
+// RUN: isa-enumeration | FileCheck %s
+// CHECK: amdgcn-amd-amdhsa--gfx600
+// CHECK: amdgcn-amd-amdhsa--gfx601
+// CHECK: amdgcn-amd-amdhsa--gfx602
+// CHECK: amdgcn-amd-amdhsa--gfx700
+// CHECK: amdgcn-amd-amdhsa--gfx701
+// CHECK: amdgcn-amd-amdhsa--gfx702
+// CHECK: amdgcn-amd-amdhsa--gfx703
+// CHECK: amdgcn-amd-amdhsa--gfx704
+// CHECK: amdgcn-amd-amdhsa--gfx705
+// CHECK: amdgcn-amd-amdhsa--gfx801
+// CHECK: amdgcn-amd-amdhsa--gfx802
+// CHECK: amdgcn-amd-amdhsa--gfx803
+// CHECK: amdgcn-amd-amdhsa--gfx805
+// CHECK: amdgcn-amd-amdhsa--gfx810
+// CHECK: amdgcn-amd-amdhsa--gfx900
+// CHECK: amdgcn-amd-amdhsa--gfx902
+// CHECK: amdgcn-amd-amdhsa--gfx904
+// CHECK: amdgcn-amd-amdhsa--gfx906
+// CHECK: amdgcn-amd-amdhsa--gfx908
+// CHECK: amdgcn-amd-amdhsa--gfx909
+// CHECK: amdgcn-amd-amdhsa--gfx90a
+// CHECK: amdgcn-amd-amdhsa--gfx90c
+// CHECK: amdgcn-amd-amdhsa--gfx942
+// CHECK: amdgcn-amd-amdhsa--gfx950
+// CHECK: amdgcn-amd-amdhsa--gfx1010
+// CHECK: amdgcn-amd-amdhsa--gfx1011
+// CHECK: amdgcn-amd-amdhsa--gfx1012
+// CHECK: amdgcn-amd-amdhsa--gfx1013
+// CHECK: amdgcn-amd-amdhsa--gfx1030
+// CHECK: amdgcn-amd-amdhsa--gfx1031
+// CHECK: amdgcn-amd-amdhsa--gfx1032
+// CHECK: amdgcn-amd-amdhsa--gfx1033
+// CHECK: amdgcn-amd-amdhsa--gfx1034
+// CHECK: amdgcn-amd-amdhsa--gfx1035
+// CHECK: amdgcn-amd-amdhsa--gfx1036
+// CHECK: amdgcn-amd-amdhsa--gfx1100
+// CHECK: amdgcn-amd-amdhsa--gfx1101
+// CHECK: amdgcn-amd-amdhsa--gfx1102
+// CHECK: amdgcn-amd-amdhsa--gfx1103
+// CHECK: amdgcn-amd-amdhsa--gfx1150
+// CHECK: amdgcn-amd-amdhsa--gfx1151
+// CHECK: amdgcn-amd-amdhsa--gfx1152
+// CHECK: amdgcn-amd-amdhsa--gfx1153
+// CHECK: amdgcn-amd-amdhsa--gfx1170
+// CHECK: amdgcn-amd-amdhsa--gfx1171
+// CHECK: amdgcn-amd-amdhsa--gfx1172
+// CHECK: amdgcn-amd-amdhsa--gfx1200
+// CHECK: amdgcn-amd-amdhsa--gfx1201
+// CHECK: amdgcn-amd-amdhsa--gfx1250
+// CHECK: amdgcn-amd-amdhsa--gfx1251
+
+// CHECK: amdgcn-amd-amdhsa--gfx9-generic
+// CHECK: amdgcn-amd-amdhsa--gfx9-4-generic
+// CHECK: amdgcn-amd-amdhsa--gfx10-1-generic
+// CHECK: amdgcn-amd-amdhsa--gfx10-3-generic
+// CHECK: amdgcn-amd-amdhsa--gfx11-generic
+// CHECK: amdgcn-amd-amdhsa--gfx12-generic
