+++ amd/comgr/test/CMakeLists.txt
@@ -231,7 +231,6 @@ add_comgr_test(compile_source_with_device_libs_to_bc_with_vfs_test c)
 add_comgr_test(assemble_test c)
 add_comgr_test(link_test c)
 add_comgr_test(isa_name_parsing_test c)
-add_comgr_test(get_data_isa_name_test c)
 add_comgr_test(include_subdirectory_test c)
 add_comgr_test(demangle_test c)
 add_comgr_test(fail_to_build_driver c)
